• LearningAuxiliary   - renamed MappingProjections for errors_signals of LearningMechanisms
