+++ Scripts/Examples/Rumelhart Semantic Network.py
@@ -60,4 +60,4 @@ S = pnl.System(processes=[rep_hidden_proc, rel_hidden_proc, rel_rep_proc, rel_pr
 # S.show_graph(show_dimensions=True)
 
 # Shows all of the learning components:
-# S.show_graph(show_learning=pnl.ALL)
+S.show_graph(show_learning=pnl.ALL)
