+++ psyneulink/components/mechanisms/adaptive/learning/learningauxilliary.py
@@ -543,19 +543,6 @@ def _instantiate_learning_components(learning_projection, context=None):
                                                                   WEIGHT: -1},
                                                           target={NAME: TARGET,
                                                                   VARIABLE: target_input},
-                                                          # input_states=[sample_input, target_input],
-                                                          # FOR TESTING: ALTERNATIVE specifications of input_states arg:
-                                                          # input_states=[(sample_input, FULL_CONNECTIVITY_MATRIX),
-                                                          # input_states=[(sample_input, RANDOM_CONNECTIVITY_MATRIX),
-                                                          #               target_input],
-                                                          # input_states=[{NAME:SAMPLE,
-                                                          #                VARIABLE:sample_input,
-                                                          #                WEIGHT:-1
-                                                          #                },
-                                                          #               {NAME:TARGET,
-                                                          #                VARIABLE:target_input,
-                                                          #                # WEIGHT:1
-                                                          #                }],
                                                           name="{} {}".format(lc.activation_mech.name,
                                                                               COMPARATOR_MECHANISM),
                                                           context=context)
@@ -623,7 +610,7 @@ def _instantiate_learning_components(learning_projection, context=None):
                                            function=learning_function,
                                            # learning_signals=[lc.activation_mech_projection],
                                            learning_signals=[learning_projection],
-                                           name=lc.activation_mech_projection.name + " " + LEARNING_MECHANISM,
+                                           name=LEARNING_MECHANISM + " for " + lc.activation_mech_projection.name,
                                            context=context)
 
     # IMPLEMENTATION NOTE:
@@ -645,13 +632,48 @@ def _instantiate_learning_components(learning_projection, context=None):
                       name = lc.activation_mech_output.owner.name + ' to ' + ACTIVATION_OUTPUT,
                       context=context)
 
-    # IMPLEMENTATION NOTE: [6/4/17] THIS IS NOW DONE IN LearningMechanism._instantiate_learning_signal
-    # # Assign learning_mechanism as sender of learning_projection and return
-    # # Note: learning_projection still has to be assigned to the learning_mechanism's outputState;
-    # #       however, this requires that it's variable be assigned (which occurs in the rest of its
-    # #       _instantiate_sender method, from which this was called) and that its value be assigned
-    # #       (which occurs in its _instantiate_function method).
-    # learning_projection.sender = learning_mechanism.output_state
+
+def _instantiate_error_signal_projection(sender, receiver):
+    """Instantiate a MappingProjection to carry an error_signal to a LearningMechanism
+
+    Can take as the sender an `ObjectiveMechanism` or a `LearningMechanism`.
+    If the sender is an ObjectiveMechanism, uses its `primary OutputState <OutputState_Primary>`.
+    If the sender is a LearningMechanism, uses its `ERROR_SIGNAL <LearningMechanism.output_states>` OutputState.
+    The receiver must be a LearningMechanism; its `ERROR_SIGNAL <LearningMechanism.input_states>` InputState is used.
+    Uses and IDENTITY_MATRIX for the MappingProjection, so requires that the sender be the same length as the receiver.
+
+    """
+    from psyneulink.components.projections.pathway.mappingprojection import MappingProjection
+
+    if isinstance(sender, ObjectiveMechanism):
+        sender = sender.output_states[OUTCOME]
+    elif isinstance(sender, LearningMechanism):
+        sender = sender.output_states[ERROR_SIGNAL]
+    else:
+        raise LearningAuxilliaryError("Sender of the error signal Projection {} "
+                                     "must be either an ObjectiveMechanism or "
+                                     "a LearningMechanism".format(sender))
+
+    if isinstance(receiver, LearningMechanism):
+        receiver = receiver.input_states[ERROR_SIGNAL]
+    else:
+        raise LearningAuxilliaryError("Receiver of the error signal Projection "
+                                     "{} must be a LearningMechanism".format(receiver))
+
+    if len(sender.value) != len(receiver.value):
+        raise LearningAuxilliaryError("The length of the OutputState ({}) for "
+                                     "the sender ({}) of the error signal "
+                                     "Projection does not match the length of "
+                                     "the InputState ({}) for the receiver "
+                                     "({})".format(len(sender.value),
+                                                   sender.owner.name,
+                                                   len(receiver.value),
+                                                   receiver.owner.name))
+
+    return MappingProjection(sender=sender,
+                             receiver=receiver,
+                             matrix=IDENTITY_MATRIX,
+                             name=sender.owner.name + ' to ' + ERROR_SIGNAL)
 
 
 @tc.typecheck
