+++ psyneulink/components/mechanisms/adaptive/learning/learningmechanism.py
@@ -1015,6 +1015,8 @@ class LearningMechanism(AdaptiveMechanism_Base):
         Also determines and assigns `error_matrices` from the `error_sources`, identified as the matrix for the
             Projection with which each error_source is associated.
         """
+        from psyneulink.components.mechanisms.adaptive.learning.learningauxilliary \
+            import _instantiate_error_signal_projection
 
         super()._instantiate_attributes_before_function(context=context)
 
@@ -1201,46 +1203,3 @@ class LearningMechanism(AdaptiveMechanism_Base):
     @property
     def learned_projections(self):
         return [lp.receiver.owner for ls in self.learning_signals for lp in ls.efferents]
-
-# IMPLEMENTATION NOTE:  THIS SHOULD BE MOVED TO COMPOSITION ONCE THAT IS IMPLEMENTED
-def _instantiate_error_signal_projection(sender, receiver):
-    """Instantiate a MappingProjection to carry an error_signal to a LearningMechanism
-
-    Can take as the sender an `ObjectiveMechanism` or a `LearningMechanism`.
-    If the sender is an ObjectiveMechanism, uses its `primary OutputState <OutputState_Primary>`.
-    If the sender is a LearningMechanism, uses its `ERROR_SIGNAL <LearningMechanism.output_states>` OutputState.
-    The receiver must be a LearningMechanism; its `ERROR_SIGNAL <LearningMechanism.input_states>` InputState is used.
-    Uses and IDENTITY_MATRIX for the MappingProjection, so requires that the sender be the same length as the receiver.
-
-    """
-    from psyneulink.components.projections.pathway.mappingprojection import MappingProjection
-
-    if isinstance(sender, ObjectiveMechanism):
-        sender = sender.output_states[OUTCOME]
-    elif isinstance(sender, LearningMechanism):
-        sender = sender.output_states[ERROR_SIGNAL]
-    else:
-        raise LearningMechanismError("Sender of the error signal Projection {} "
-                                     "must be either an ObjectiveMechanism or "
-                                     "a LearningMechanism".format(sender))
-
-    if isinstance(receiver, LearningMechanism):
-        receiver = receiver.input_states[ERROR_SIGNAL]
-    else:
-        raise LearningMechanismError("Receiver of the error signal Projection "
-                                     "{} must be a LearningMechanism".format(receiver))
-
-    if len(sender.value) != len(receiver.value):
-        raise LearningMechanismError("The length of the OutputState ({}) for "
-                                     "the sender ({}) of the error signal "
-                                     "Projection does not match the length of "
-                                     "the InputState ({}) for the receiver "
-                                     "({})".format(len(sender.value),
-                                                   sender.owner.name,
-                                                   len(receiver.value),
-                                                   receiver.owner.name))
-
-    return MappingProjection(sender=sender,
-                             receiver=receiver,
-                             matrix=IDENTITY_MATRIX,
-                             name=sender.owner.name + ' ' + OUTCOME)
